all: smoother configurations repository server url updating (fixes #15754) (#15658)
