+++ app/build.gradle
@@ -10,8 +10,8 @@ android {
         applicationId "org.ole.planet.myplanet"
         minSdk = 26
         targetSdk = 36
-        versionCode = 6492
-        versionName = "0.64.92"
+        versionCode = 6493
+        versionName = "0.64.93"
         ndkVersion = '26.3.11579264'
         vectorDrawables.useSupportLibrary = true
     }
