+++ app/src/main/java/org/ole/planet/myplanet/ui/courses/ProgressViewModel.kt
@@ -9,12 +9,12 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import org.ole.planet.myplanet.repository.ProgressRepository
-import org.ole.planet.myplanet.services.UserSessionManager
+import org.ole.planet.myplanet.repository.UserRepository
 
 @HiltViewModel
 class ProgressViewModel @Inject constructor(
     private val progressRepository: ProgressRepository,
-    private val userSessionManager: UserSessionManager
+    private val userRepository: UserRepository
 ) : ViewModel() {
 
     private val _courseData = MutableStateFlow<JsonArray?>(null)
@@ -22,7 +22,7 @@ class ProgressViewModel @Inject constructor(
 
     fun loadCourseData() {
         viewModelScope.launch {
-            val user = userSessionManager.getUserModel()
+            val user = userRepository.getUserModel()
             _courseData.value = progressRepository.fetchCourseData(user?.id)
         }
     }
