+++ app/src/test/java/org/ole/planet/myplanet/ui/courses/CourseProgressViewModelTest.kt
@@ -11,7 +11,7 @@ import org.junit.Test
 import org.ole.planet.myplanet.model.CourseProgressData
 import org.ole.planet.myplanet.model.UserEntity
 import org.ole.planet.myplanet.repository.CoursesRepository
-import org.ole.planet.myplanet.services.UserSessionManager
+import org.ole.planet.myplanet.repository.UserRepository
 import org.ole.planet.myplanet.utils.MainDispatcherRule
 
 class CourseProgressViewModelTest {
@@ -20,8 +20,8 @@ class CourseProgressViewModelTest {
     val mainDispatcherRule = MainDispatcherRule()
 
     private val coursesRepository = mockk<CoursesRepository>()
-    private val userSessionManager = mockk<UserSessionManager>()
-    private val viewModel = CourseProgressViewModel(coursesRepository, userSessionManager)
+    private val userRepository = mockk<UserRepository>()
+    private val viewModel = CourseProgressViewModel(coursesRepository, userRepository)
 
     @Test
     fun `courseProgress value is null before any call`() {
@@ -34,7 +34,7 @@ class CourseProgressViewModelTest {
         val userId = "userId"
         val user = UserEntity()
         user._id = userId
-        coEvery { userSessionManager.getUserModel() } returns user
+        coEvery { userRepository.getUserModel() } returns user
 
         val expectedProgressData = mockk<CourseProgressData>()
         coEvery { coursesRepository.getCourseProgress(courseId, userId) } returns expectedProgressData
@@ -50,7 +50,7 @@ class CourseProgressViewModelTest {
         val userId = "userId"
         val user = UserEntity()
         user._id = userId
-        coEvery { userSessionManager.getUserModel() } returns user
+        coEvery { userRepository.getUserModel() } returns user
 
         val expectedProgressData = mockk<CourseProgressData>()
         coEvery { coursesRepository.getCourseProgress(courseId, userId) } returns expectedProgressData
